requirements.txt
asgiref==3.7.2
Django==4.2.6
django-environ==0.11.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
protobuf==3.20.3
PyJWT==2.8.0
PyMySQL==1.1.0
pytz==2023.3.post1
setuptools==68.2.2
sqlparse==0.4.4
tzdata==2023.3
